Added sentence generation engine
The sentence generation engine, which I call SentenceGen, is a simple Grammar machine which takes strings in a set format and uses them to generate whole sentences. The current data repo to build from is a little lacking at the moment. That will come in later commits, but I might not need to actually change the code at this point to get new sentences.